refactor(xapiri): extract renderAboutTab → tab_about.go
Mechanical move per ADR 0014. No behaviour change.
