+++ internal/ui/xapiri/dashboard.go
@@ -27,7 +27,6 @@ import (
 	"log/slog"
 	"os"
 	"os/exec"
-	"runtime/debug"
 	"sort"
 	"strconv"
 	"strings"
@@ -2438,42 +2437,6 @@ func (m dashModel) renderHelpTab(w, h int) string {
 		"  " + stAccent.Render("ctrl+alt+↑/↓") + "     resize terminal pane",
 		"  " + stAccent.Render("ctrl+s") + "           save config and continue",
 		"  " + stAccent.Render("esc / q") + "          abort (no changes written)",
-	}
-	for len(lines) < h {
-		lines = append(lines, "")
-	}
-	return strings.Join(lines[:min(len(lines), h)], "\n")
-}
-
-// ─── about tab ───────────────────────────────────────────────────────────────
-
-func (m dashModel) renderAboutTab(w, h int) string {
-	version := "unknown"
-	commit := "unknown"
-	if info, ok := debug.ReadBuildInfo(); ok {
-		version = info.Main.Version
-		for _, s := range info.Settings {
-			if s.Key == "vcs.revision" && len(s.Value) >= 7 {
-				commit = s.Value[:7]
-			}
-		}
-	}
-
-	lines := []string{
-		stHdr.Render(" About yage xapiri"),
-		stMuted.Render(strings.Repeat("─", w)),
-		"",
-		"  " + stBold.Render("yage") + "  — Yet Another GitOps Engine",
-		"",
-		"  " + stMuted.Render("version:") + "  " + stAccent.Render(version),
-		"  " + stMuted.Render("commit: ") + "  " + stAccent.Render(commit),
-		"",
-		"  " + stMuted.Render("license:") + "  Apache-2.0",
-		"  " + stMuted.Render("project:") + "  https://github.com/lpasquali/yage",
-		"",
-		stMuted.Render("  xapiri are sacred spirits in the Yanomami people's cosmology."),
-		stMuted.Render("  yage runs xapiri to get help from the spirits to create a"),
-		stMuted.Render("  visionary deployment."),
 	}
 	for len(lines) < h {
 		lines = append(lines, "")
